+++ R/ExperimentRegistry.R
@@ -29,7 +29,7 @@
 #' summarizeExperiments(reg = tmp)
 #' summarizeExperiments(reg = tmp, by = c("problem", "algorithm", "n"))
 #' ids = findExperiments(prob.pars = (n == 50), reg = tmp)
-#' print(flatten(getJobPars(ids, reg = tmp)))
+#' print(unwrap(getJobPars(ids, reg = tmp)))
 #'
 #' # Submit jobs
 #' submitJobs(reg = tmp)
@@ -42,8 +42,8 @@
 #' # Join info table with all results and calculate mean of results
 #' # grouped by n and algorithm
 #' ids = findDone(reg = tmp)
-#' pars = flatten(getJobPars(ids, reg = tmp))
-#' results = flatten(reduceResultsDataTable(ids, fun = function(res) list(res = res), reg = tmp))
+#' pars = unwrap(getJobPars(ids, reg = tmp))
+#' results = unwrap(reduceResultsDataTable(ids, fun = function(res) list(res = res), reg = tmp))
 #' tab = ljoin(pars, results)
 #' tab[, list(mres = mean(res)), by = c("n", "algorithm")]
 makeExperimentRegistry = function(file.dir = "registry", work.dir = getwd(), conf.file = findConfFile(), packages = character(0L), namespaces = character(0L),
@@ -81,9 +81,4 @@ print.ExperimentRegistry = function(x, ...) {
   catf("  Algorithms: %i", nlevels(x$defs$algorithm))
   catf("  Seed      : %i", x$seed)
   catf("  Writeable : %s", x$writeable)
-}
-
-assertExperimentRegistry = function(reg, writeable = FALSE, sync = FALSE, running.ok = TRUE) {
-  assertClass(reg, "ExperimentRegistry")
-  assertRegistry(reg, writeable = writeable, sync = sync, running.ok = running.ok)
 }