+++ R/summarizeExperiments.R
@@ -13,10 +13,10 @@
 #' @export
 #' @family Experiment
 summarizeExperiments = function(ids = NULL, by = c("problem", "algorithm"), reg = getDefaultRegistry()) {
-  assertExperimentRegistry(reg)
+  assertRegistry(reg, class = "ExperimentRegistry")
   assertCharacter(by, any.missing = FALSE, min.chars = 1L, min.len = 1L, unique = TRUE)
   tab = getJobPars(ids = ids, reg = reg)
   if (!setequal(by, c("problem", "algorithm")))
-    tab = flatten(tab)
+    tab = unwrap(tab)
   tab[, list(.count = .N), by = by]
 }