+++ tests/testthat/test_estimateRuntimes.R
@@ -7,7 +7,7 @@ test_that("estimateRuntimes", {
   addAlgorithm(name = "ncol", function(instance, ...) ncol(instance), reg = reg)
   addExperiments(algo.designs = list(nrow = CJ(x = 1:50, y = letters[1:5])), reg = reg)
   addExperiments(algo.designs = list(ncol = CJ(x = 1:50, y = letters[1:5])), reg = reg)
-  tab = flatten(getJobPars(reg = reg))
+  tab = unwrap(getJobPars(reg = reg))
 
   ids = tab[, .SD[sample(nrow(.SD), 4)], by = c("problem", "algorithm", "y")]
   setkeyv(ids, "job.id")
@@ -33,6 +33,6 @@ test_that("estimateRuntimes", {
   expect_true(all(res[y == "a", t] > res[y %in% c("c", "d", "e"), t]))
 
   # remaining is suppressed if nothing more to submit, no error
-  res = estimateRuntimes(flatten(getJobPars(findDone(reg = reg), reg = reg)), reg = reg)
+  res = estimateRuntimes(unwrap(getJobPars(findDone(reg = reg), reg = reg)), reg = reg)
   expect_output(print(res, n = 2))
 })